renamed eso dungeon table files
