+++ src/GM-Manager-Flask/test/integration_tests/test_eso_api.py
@@ -1,7 +1,7 @@
 import requests
 import constants
 from modes.eso import constants as eso_constants
-from database.query_eso_table import QueryEsoTable
+from database.query_eso_dungeon_table import QueryEsoTable
 
 
 class TestEsoApi:
